docs/python/requirements.txt
Jinja2==3.1.4
MarkupSafe==2.1.5
PyYAML==6.0.1